Move user_dict from DataSet to DataObject
