+++ pyvista/core/dataobject.py
@@ -1,6 +1,7 @@
 """Attributes common to PolyData and Grid Objects."""
 
 from abc import abstractmethod
+from collections import UserDict
 import collections.abc
 from pathlib import Path
 from typing import Any, ClassVar, DefaultDict, Dict, Optional, Type, Union
@@ -12,7 +13,8 @@ import pyvista
 from . import _vtk_core as _vtk
 from ._typing_core import NumpyArray
 from .datasetattributes import DataSetAttributes
-from .utilities.arrays import FieldAssociation
+from .pyvista_ndarray import pyvista_ndarray
+from .utilities.arrays import FieldAssociation, _JSONValueType, _SerializedDictArray
 from .utilities.fileio import read, set_vtkwriter_mode
 from .utilities.helpers import wrap
 from .utilities.misc import abstract_class
@@ -460,6 +462,150 @@ class DataObject:
             raise NotImplementedError(f'`{type(self)}` does not support field data')
 
         self.field_data.clear()
+
+    @property
+    def user_dict(self) -> _SerializedDictArray:
+        """Set or return a user-specified data dictionary.
+
+        The dictionary is stored as a JSON-serialized string as part of the mesh's
+        field data. Unlike regular field data, which requires values to be stored
+        as an array, the user dict provides a mapping for scalar values.
+
+        Since the user dict is stored as field data, it is automatically saved
+        with the mesh when it is saved in a compatible file format (e.g. ``'.vtk'``).
+        Any saved metadata is automatically de-serialized by PyVista whenever
+        the user dict is accessed again. Since the data is stored as JSON, it
+        may also be easily retrieved or read by other programs.
+
+        Any JSON-serializable values are permitted by the user dict, i.e. values
+        can have type ``dict``, ``list``, ``tuple``, ``str``, ``int``, ``float``,
+        ``bool``, or ``None``. Storing NumPy arrays is not directly supported, but
+        these may be cast beforehand to a supported type, e.g. by calling ``tolist()``
+        on the array.
+
+        To completely remove the user dict string from the dataset's field data,
+        set its value to ``None``.
+
+        .. note::
+
+            The user dict is a convenience property and is intended for metadata storage.
+            It has an inefficient dictionary implementation and should only be used to
+            store a small number of infrequently-accessed keys with relatively small
+            values. It should not be used to store frequently accessed array data
+            with many entries (a regular field data array should be used instead).
+
+        .. warning::
+
+            Field data is typically passed-through by dataset filters, and therefore
+            the user dict's items can generally be expected to persist and remain
+            unchanged in the output of filtering methods. However, this behavior is
+            not guaranteed, as it's possible that some filters may modify or clear
+            field data. Use with caution.
+
+        Returns
+        -------
+        UserDict
+            JSON-serialized dict-like object which is subclassed from :py:class:`collections.UserDict`.
+
+        Examples
+        --------
+        Load a mesh.
+
+        >>> import pyvista as pv
+        >>> from pyvista import examples
+        >>> mesh = examples.load_ant()
+
+        Add data to the user dict. The contents are serialized as JSON.
+
+        >>> mesh.user_dict['name'] = 'ant'
+        >>> mesh.user_dict
+        {"name": "ant"}
+
+        Alternatively, set the user dict from an existing dict.
+
+        >>> mesh.user_dict = dict(name='ant')
+
+        The user dict can be updated like a regular dict.
+
+        >>> mesh.user_dict.update(
+        ...     {
+        ...         'num_legs': 6,
+        ...         'body_parts': ['head', 'thorax', 'abdomen'],
+        ...     }
+        ... )
+        >>> mesh.user_dict
+        {"name": "ant", "num_legs": 6, "body_parts": ["head", "thorax", "abdomen"]}
+
+        Data in the user dict is stored as field data.
+
+        >>> mesh.field_data
+        pyvista DataSetAttributes
+        Association     : NONE
+        Contains arrays :
+            _PYVISTA_USER_DICT      str        "{"name": "ant",..."
+
+        Since it's field data, the user dict can be saved to file along with the
+        mesh and retrieved later.
+
+        >>> mesh.save('ant.vtk')
+        >>> mesh_from_file = pv.read('ant.vtk')
+        >>> mesh_from_file.user_dict
+        {"name": "ant", "num_legs": 6, "body_parts": ["head", "thorax", "abdomen"]}
+
+        """
+        self._config_user_dict()
+        return self._user_dict
+
+    @user_dict.setter
+    def user_dict(
+        self,
+        dict_: Union[dict[str, _JSONValueType], UserDict[str, _JSONValueType]],
+    ):  # numpydoc ignore=GL08
+
+        # Setting None removes the field data array
+        if dict_ is None and '_PYVISTA_USER_DICT' in self.field_data.keys():
+            del self.field_data['_PYVISTA_USER_DICT']
+            return
+
+        self._config_user_dict()
+        if isinstance(dict_, dict):
+            self._user_dict.data = dict_
+        elif isinstance(dict_, UserDict):
+            self._user_dict.data = dict_.data
+        else:
+            raise TypeError(
+                f'User dict can only be set with type {dict} or {UserDict}.\nGot {type(dict_)} instead.',
+            )
+
+    def _config_user_dict(self):
+        """Init serialized dict array and ensure it is added to field_data."""
+        field_name = '_PYVISTA_USER_DICT'
+        field_data = self.field_data
+
+        if not hasattr(self, '_user_dict'):
+            # Init
+            self._user_dict = _SerializedDictArray()
+
+        if field_name in field_data.keys():
+            if isinstance(array := field_data[field_name], pyvista_ndarray):
+                # When loaded from file, field will be cast as pyvista ndarray
+                # Convert to string and initialize new user dict object from it
+                self._user_dict = _SerializedDictArray(''.join(array))
+            elif isinstance(array, str) and repr(self._user_dict) != array:
+                # Filters may update the field data block separately, e.g.
+                # when copying field data, so we need to capture the new
+                # string and re-init
+                self._user_dict = _SerializedDictArray(array)
+            else:
+                # User dict is correctly configured, do nothing
+                return
+
+        # Set field data array directly instead of calling 'set_array'
+        # This skips the call to '_prepare_array' which will otherwise
+        # do all kinds of casting/conversions and mangle this array
+        self._user_dict.SetName(field_name)
+        field_data.VTKObject.AddArray(self._user_dict)
+        field_data.VTKObject.Modified()
 
     @property
     def memory_address(self) -> str:  # numpydoc ignore=RT01
